# Add idle timeout to queue proxy