--- a/cmd/queue/main.go
+++ b/cmd/queue/main.go
@@ -73,13 +73,14 @@ var (
 )
 
 type config struct {
-	ContainerConcurrency     int    `split_words:"true" required:"true"`
-	QueueServingPort         string `split_words:"true" required:"true"`
-	UserPort                 string `split_words:"true" required:"true"`
-	RevisionTimeoutSeconds   int    `split_words:"true" required:"true"`
-	ServingReadinessProbe    string `split_words:"true" required:"true"`
-	EnableProfiling          bool   `split_words:"true"` // optional
-	EnableHTTP2AutoDetection bool   `split_words:"true"` // optional
+	ContainerConcurrency       int    `split_words:"true" required:"true"`
+	QueueServingPort           string `split_words:"true" required:"true"`
+	UserPort                   string `split_words:"true" required:"true"`
+	RevisionTimeoutSeconds     int    `split_words:"true" required:"true"`
+	ServingReadinessProbe      string `split_words:"true" required:"true"`
+	RevisionIdleTimeoutSeconds int    `split_words:"true" default:"-1"` // optional
+	EnableProfiling            bool   `split_words:"true"`              // optional
+	EnableHTTP2AutoDetection   bool   `split_words:"true"`              // optional
 
 	// Logging configuration
 	ServingLoggingConfig         string `split_words:"true" required:"true"`
@@ -294,7 +295,8 @@ func buildServer(ctx context.Context, env config, healthState *health.State, rp 
 	metricsSupported := supportsMetrics(ctx, logger, env)
 	tracingEnabled := env.TracingConfigBackend != tracingconfig.None
 	concurrencyStateEnabled := env.ConcurrencyStateEndpoint != ""
-	timeout := time.Duration(env.RevisionTimeoutSeconds) * time.Second
+	firstByteTimeout := time.Duration(env.RevisionTimeoutSeconds) * time.Second
+	idleTimeout := time.Duration(env.RevisionIdleTimeoutSeconds) * time.Second
 
 	// Create queue handler chain.
 	// Note: innermost handlers are specified first, ie. the last handler in the chain will be executed first.
@@ -308,7 +310,7 @@ func buildServer(ctx context.Context, env config, healthState *health.State, rp 
 	}
 	composedHandler = queue.ProxyHandler(breaker, stats, tracingEnabled, composedHandler)
 	composedHandler = queue.ForwardedShimHandler(composedHandler)
-	composedHandler = handler.NewTimeToFirstByteTimeoutHandler(composedHandler, "request timeout", timeout)
+	composedHandler = handler.NewTimeoutHandler(composedHandler, "request timeout", firstByteTimeout, idleTimeout)
 
 	if metricsSupported {
 		composedHandler = requestMetricsHandler(logger, composedHandler, env)