--- a/pkg/http/handler/timeout.go
+++ b/pkg/http/handler/timeout.go
@@ -80,18 +80,18 @@ func (h *timeoutHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 
 	firstByteTimeout := getTimer(h.firstByteTimeout)
 	var firstByteTimeoutDrained bool
-
-	var idleTimeout *optionalTimer
-	if h.idleTimeout < 0 {
-		idleTimeout = newOptionalTimer(nil)
-	} else {
-		idleTimeout = newOptionalTimer(getTimer(h.idleTimeout))
-	}
-
 	defer func() {
 		putTimer(firstByteTimeout, firstByteTimeoutDrained)
-		putOptionalTimer(idleTimeout)
 	}()
+
+	var idleTimeout *time.Timer
+	var idleTimeoutDrained bool
+	if h.idleTimeout > 0 {
+		idleTimeout = getTimer(h.idleTimeout)
+		defer func() {
+			putTimer(idleTimeout, idleTimeoutDrained)
+		}()
+	}
 
 	for {
 		select {
@@ -105,10 +105,10 @@ func (h *timeoutHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 			if tw.tryFirstByteTimeoutAndWriteError(h.body) {
 				return
 			}
-		case <-idleTimeout.C():
+		case <-timerChannel(idleTimeout):
 			timedOut, timeToNextTimeout := tw.tryIdleTimeoutAndWriteError(time.Now(), h.idleTimeout, h.body)
 			if timedOut {
-				idleTimeout.SetDrained(true)
+				idleTimeoutDrained = true
 				return
 			}
 			idleTimeout.Reset(timeToNextTimeout)
@@ -116,50 +116,15 @@ func (h *timeoutHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	}
 }
 
-// optionalTimer provides a wrapper for time.Timer. When given a not-nil Timer pointer, it
-// passes operations to the Timer. When given a nil Timer pointer, it fakes the behavior of
-// a timer with a no-op channel and no-op operations
-type optionalTimer struct {
-	timer    *time.Timer
-	drained  bool
-	disabled bool
-}
+var never = make(<-chan time.Time)
 
-func newOptionalTimer(timer *time.Timer) *optionalTimer {
-	return &optionalTimer{
-		disabled: timer == nil,
-		drained:  timer == nil,
-		timer:    timer,
+// timerChannel returns the channel of the given timer or a channel that never fires
+// if the timer is nil.
+func timerChannel(timer *time.Timer) <-chan time.Time {
+	if timer != nil {
+		return timer.C
 	}
-}
-
-func (ot *optionalTimer) C() <-chan time.Time {
-	if ot.disabled {
-		return make(<-chan time.Time)
-	}
-	return ot.timer.C
-}
-
-func (ot *optionalTimer) Timer() (exists bool, timer *time.Timer) {
-	if ot.disabled {
-		return false, nil
-	}
-	return true, ot.timer
-}
-
-func (ot *optionalTimer) Reset(d time.Duration) bool {
-	if ot.disabled {
-		return false
-	}
-	return ot.timer.Reset(d)
-}
-
-func (ot *optionalTimer) Drained() bool {
-	return ot.drained
-}
-
-func (ot *optionalTimer) SetDrained(drained bool) {
-	ot.drained = drained
+	return never
 }
 
 // timeoutWriter is a wrapper around an http.ResponseWriter. It guards
@@ -286,10 +251,4 @@ func putTimer(t *time.Timer, alreadyDrained bool) {
 		<-t.C
 	}
 	timerPool.Put(t)
-}
-
-func putOptionalTimer(t *optionalTimer) {
-	if exists, timer := t.Timer(); exists {
-		putTimer(timer, t.Drained())
-	}
 }