--- a/pkg/http/handler/timeout_test.go
+++ b/pkg/http/handler/timeout_test.go
@@ -31,7 +31,7 @@ func TestTimeoutWriterAllowsForAdditionalWritesBeforeTimeout(t *testing.T) {
 	handler := &timeoutWriter{w: recorder}
 	handler.WriteHeader(http.StatusOK)
 	handler.tryFirstByteTimeoutAndWriteError("error")
-	handler.tryIdleTimeoutAndWriteError(10*time.Second, "error")
+	handler.tryIdleTimeoutAndWriteError(time.Now(), 10*time.Second, "error")
 	if _, err := io.WriteString(handler, "test"); err != nil {
 		t.Fatalf("handler.Write() = %v, want no error", err)
 	}
@@ -71,6 +71,35 @@ func TestTimeoutWriterErrorsWriteAfterTimeout(t *testing.T) {
 	handler.timeoutAndWriteError("error")
 	if _, err := handler.Write([]byte("hello")); !errors.Is(err, http.ErrHandlerTimeout) {
 		t.Errorf("ErrHandlerTimeout got %v, want: %s", err, http.ErrHandlerTimeout)
+	}
+}
+
+func TestDisabledOptionalTimerHasNoTimerAllocated(t *testing.T) {
+	optTimer := newOptionalTimer(nil)
+	wantExists := false
+	var wantTimer *time.Timer
+
+	if gotExists, gotTimer := optTimer.Timer(); gotExists != wantExists || gotTimer != wantTimer {
+		t.Errorf("For disabled optionalTimer, got timer exists: %t, want timer exists: %t, got timer: %v, want: %v", gotExists, wantExists, gotTimer, wantTimer)
+	}
+}
+
+func TestEnabledOptionalTimerHasTimerAllocated(t *testing.T) {
+	optTimer := newOptionalTimer(time.NewTimer(1))
+	gotExists, _ := optTimer.Timer()
+
+	wantExists := true
+	if gotExists != wantExists {
+		t.Errorf("For enabled optionalTimer, got timer exists: %t, want timer exists: %t", gotExists, wantExists)
+	}
+}
+
+func TestEnabledOptionalTimerReturnsChannelOfInnerTimer(t *testing.T) {
+	optTimer := newOptionalTimer(time.NewTimer(1))
+	_, innerTimer := optTimer.Timer()
+
+	if innerTimer.C != optTimer.C() {
+		t.Errorf("Enabled optionalTimer.C() should return the channel of its inner timer")
 	}
 }