Fixes https://github.com/knative/serving/issues/10852

<!-- Please include the 'why' behind your changes if no issue exists -->
## Proposed Changes

* Add idle timeout to queue proxy
* Refactor timeout handler tests

benchmark before and after the change (needed to do some manual renaming in the benchmark output because benchmark test name changed)
```
benchmark                                 old ns/op     new ns/op     delta
BenchmarkTimeoutHandler/sequential-12     1311          2180          +66.29%
BenchmarkTimeoutHandler/parallel-12       182           274           +50.52%

benchmark                                 old allocs     new allocs     delta
BenchmarkTimeoutHandler/sequential-12     6              6              +0.00%
BenchmarkTimeoutHandler/parallel-12       6              6              +0.00%

benchmark                                 old bytes     new bytes     delta
BenchmarkTimeoutHandler/sequential-12     647           673           +4.02%
BenchmarkTimeoutHandler/parallel-12       645           666           +3.26%
```

**Release Note**

<!-- Enter your extended release note in the below block. If the PR requires
additional action from users switching to the new release, include the string
"action required". If no release note is required, write "NONE". -->

```release-note

```