@nealhu 2021-08-09
This is still a working progress because I still need to add logics to set idle timeout defaults, but I want to get some feedback. 
* As discussed in #10852 , instead of making a separate handler, I'm extending the existing `timeToFirstByteTimeoutHandler` to be a more generic `timeoutHandler`. That makes the code a little more complex, what do people think? Good enough or separate the handlers now?
* I also want to confirm the API,  is `RevisionIdleTimeoutSeconds` a good name?