@julz 2021-08-09
/assign @markusthoemmes for the timeout handler changes (especially since they touch the new timer pooling)